+++ Source/Network/TCPCommand/TCPRoom.cpp
@@ -80,7 +80,7 @@ namespace Online
 	*//***************************************************************************/
 	bool TCPRoomOut::Receive(size_t size)
 	{
-		uint64_t client_id = 0;
+		uint32_t client_id = 0;
 		if (m_pcontroller->GetTcpSocket()->Receive(&client_id, size) > 0)
 		{
 			if (client_id == GAME_DATA.GetClientId()) return true;
